refactor: extract browser utilities
