+++ src/prepare/runtime.ts
@@ -16,13 +16,11 @@ export function prepareRuntime(ctx: I18nNuxtContext, nuxt: Nuxt) {
 
   // for composables
   nuxt.options.alias['#i18n'] = resolver.resolve('./runtime/composables/index')
+  nuxt.options.alias['#i18n-kit'] = resolver.resolve('./runtime/kit')
   nuxt.options.alias['#internal-i18n-types'] = resolver.resolve('./types')
-  nuxt.options.alias['#i18n-kit/head'] = resolver.resolve('./runtime/kit/head')
-  nuxt.options.alias['#i18n-kit/routing'] = resolver.resolve('./runtime/kit/routing')
   nuxt.options.build.transpile.push('#i18n')
+  nuxt.options.build.transpile.push('#i18n-kit')
   nuxt.options.build.transpile.push('#internal-i18n-types')
-  nuxt.options.build.transpile.push('#i18n-kit/head')
-  nuxt.options.build.transpile.push('#i18n-kit/routing')
 
   if (nuxt.options.dev && options.hmr) {
     addVitePlugin({
