--- src/runtime/browser.ts
@@ -1,14 +1,0 @@
-/**
- * Parses locales provided from browser through `accept-language` header.
- *
- * @param input - Accept-Language header value.
- * @return An array of locale codes. Priority determined by order in array.
- */
-export function parseAcceptLanguage(input: string = ''): string[] {
-  // Example input: en-US,en;q=0.9,nb;q=0.8,no;q=0.7
-  // Contains tags separated by comma.
-  // Each tag consists of locale code (2-3 letter language code) and optionally country code
-  // after dash. Tag can also contain score after semicolon, that is assumed to match order
-  // so it's not explicitly used.
-  return input.split(',').map(tag => tag.split(';')[0])
-}
