+++ src/runtime/server/utils/locale-detector.ts
@@ -2,9 +2,8 @@ import { fetchMessages } from '../context'
 import { deepCopy } from '@intlify/shared'
 import { localeDetector } from '#internal/i18n/locale.detector.mjs'
 import { tryCookieLocale, tryHeaderLocale, tryQueryLocale } from '@intlify/h3'
-import { findBrowserLocale } from '#i18n-kit/routing'
+import { findBrowserLocale, parseAcceptLanguage } from '#i18n-kit/browser'
 import { normalizedLocales } from '#internal/i18n/options.mjs'
-import { parseAcceptLanguage } from '../../browser'
 
 import type { H3Event } from 'h3'
 import type { CoreOptions, FallbackLocale, Locale } from '@intlify/core'
